backend-project/django/class-notes/DRF_Pagination_Filtering_Search/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
